Add files via upload
The master branch has some GUI and front end loop completed, however additional parameters and instances were necessary for the creation of a (semi) working Mob class. Thus I not only created logic for Mob movement, but I altered the existing code, we had so progress can be made towards a working Mob method. Further editing is needed of player and main files in order for Mob to be fully functioning, but I refrained from doing such major changes independently and frivolously. Thus I created a second branch with edits and a Mob class that has functioning with regard to being able to track our player and move towards it (the difficulty lies in when the player begins to move, but as I said above that requires more serious editing of other files).